# bigdata-vis/api/swagger/swagger (copia).yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: Big Data Visualization API
# during dev, should point to your local machine
host: localhost:20183
...
basePath: /
...
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:
...
  /datavis/scatterplot:
    x-swagger-router-controller: scatterplot
    get:
      description: Returns information about the requested scatterplot
      operationId: scatterplot
      parameters:
        - name: file
          in: query
          description: Path to file
          required: true
          type: string
        - name: colx
          in: query
          description: The column selected for the scatterplot
          required: true
          type: string
        - name: coly
          in: query
          description: The column selected for the scatterplot
          required: true
          type: string
        - name: secx
          in: query
          description: Number of sections to divide values for the x-axis
          required: true
          type: number
        - name: secy
          in: query
          description: Number of sections to divide values for the y-axis
          required: true
          type: number
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/ScatterplotResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
...
  /datavis/piechart:
    x-swagger-router-controller: piechart
    get:
      description: Returns information about the requested piechart
      operationId: piechart
      parameters:
        - name: file
          in: query
          description: Path to file
          required: true
          type: string
        - name: colSelected
          in: query
          description: The column selected for the piechart
          required: true
          type: string
        - name: colCount
          in: query
          description: Column that count values
          required: true
          type: string
        - name: opVal
          in: query
          description: Number of operation to apply [0-Count, 1-Sum, 2-Max, 3-Min]
          required: true
          type: number
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/PiechartResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
...
definitions:
...
  ScatterplotResponse:
    properties:
      _id:
        type: string
      title:
        type: string
      titleX:
        type: string
      titleY:
        type: string
      values:
        type: array
        items:
          type: object
          properties:
            pointX:
              type: number
            pointY:
              type: number
            value:
              type: number
            valueColor:
              type: number
...
  PiechartResponse:
    properties:
      _id:
        type: string
      title:
        type: string
      colSelected:
        type: string
      colCount:
        type: string
      lvalues:
        type: array
        items:
          type: object
          properties:
            colKey:
              type: string
            colValue:
              type: number
...
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string